Copied in the PY210L9 mailroom files and deleted the PY220L4 mailroom files, since the assignment is based on expansion of the original object-oriented mailroom (not the JSON/metaclass version).
